+++ packages/fluentui/react-northstar/src/index.ts
@@ -116,6 +116,12 @@ export * from './components/Layout/Layout';
 
 export * from './components/List/List';
 export * from './components/List/ListItem';
+export * from './components/List/ListItemContent';
+export * from './components/List/ListItemContentMedia';
+export * from './components/List/ListItemEndMedia';
+export * from './components/List/ListItemHeader';
+export * from './components/List/ListItemHeaderMedia';
+export * from './components/List/ListItemMedia';
 
 export * from './components/Menu/Menu';
 export * from './components/Menu/MenuItem';
@@ -172,10 +178,6 @@ export * from './components/Toolbar/ToolbarMenuItemActiveIndicator';
 export * from './components/Toolbar/ToolbarMenuRadioGroup';
 export * from './components/Toolbar/ToolbarMenuRadioGroupWrapper';
 export * from './components/Toolbar/ToolbarRadioGroup';
-
-export * from './components/HierarchicalTree/HierarchicalTree';
-export * from './components/HierarchicalTree/HierarchicalTreeItem';
-export * from './components/HierarchicalTree/HierarchicalTreeTitle';
 
 export * from './components/Tree/Tree';
 export * from './components/Tree/TreeItem';